requirements.txt
mcp[cli]==1.9.4
python-dotenv==1.2.1
pydantic==2.12.5
httpx==0.28.1
Jinja2==3.1.6
PyYAML==6.0.2
mistune==3.1.3
